Consolidate local and remote reducers into api
There is now one api reducer.

This commit also removes a Models dependency on a Store. This produced
a smell as Models were not truly an extension of store. Models now
possess all querying methods and store functionality
