+++ .gitignore
@@ -3,3 +3,4 @@
 .npmrc
 .vimrc
 /coverage
+dist
