+++ active-redux.js
@@ -1,0 +1,9 @@
+import _Model from './model';
+export { _Model as Model };
+import _Registry from './registry';
+export { _Registry as Registry };
+import * as _Attr from './attributes';
+export { _Attr as Attr };
+
+export * from './redux/local';
+export * from './redux/remote';
