+++ example/src/configureStore.js
@@ -1,0 +1,38 @@
+import { applyMiddleware, compose, combineReducers, createStore as createReduxStore } from 'redux';
+import thunk from 'redux-thunk';
+import { reducer } from '../../dist';
+
+const createStore = (initialState = {}) => {
+  // ======================================================
+  // Middleware Configuration
+  // ======================================================
+  const middleware = [thunk]
+
+  // ======================================================
+  // Store Enhancers
+  // ======================================================
+  const enhancers = []
+  let composeEnhancers = compose
+
+  if (process.env.NODE_ENV === 'development') {
+    if (typeof window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ === 'function') {
+      composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__
+    }
+  }
+
+  // ======================================================
+  // Store Instantiation
+  // ======================================================
+  const store = createReduxStore(
+    combineReducers(reducer),
+    initialState,
+    composeEnhancers(
+      applyMiddleware(...middleware),
+      ...enhancers
+    )
+  )
+
+  return store
+}
+
+export default createStore
