+++ example/src/models/Article.js
@@ -1,0 +1,5 @@
+import { Model } from '../../dist';
+
+export default class Article extends Model {
+
+}
