+++ example/src/models/Comment.js
@@ -1,0 +1,5 @@
+import { Model } from '../../dist';
+
+export default class Comment extends Model {
+
+}
