+++ example/src/models/User.js
@@ -1,0 +1,5 @@
+import { Model } from '../../dist';
+
+export default class User extends Model {
+
+}
