+++ package.json
@@ -1,10 +1,14 @@
 {
-  "name": "ch-json-api",
+  "name": "@coverhound/active-redux",
   "version": "1.0.0",
-  "main": "index.js",
+  "main": "dist/index.js",
   "license": "MIT",
   "scripts": {
+    "babel": "babel src --out-dir dist",
+    "build": "rm -rf dist && npm run babel",
+    "dev": "npm run babel -- -w",
     "lint": "eslint .",
+    "postinstall": "npm run build",
     "test": "jest",
     "test:ci": "jest --ci",
     "test:coverage": "jest --coverage",
@@ -19,6 +23,7 @@
   },
   "devDependencies": {
     "@coverhound/eslint-config-coverhound": "^1.0.1",
+    "babel-cli": "^6.24.1",
     "babel-jest": "^20.0.3",
     "babel-polyfill": "^6.23.0",
     "babel-preset-es2015": "^6.24.1",
@@ -26,7 +31,8 @@
     "eslint": "^3.19.0",
     "eslint-plugin-jest": "^20.0.3",
     "jest": "^20.0.3",
-    "jest-cli": "^20.0.3"
+    "jest-cli": "^20.0.3",
+    "redux": "^3.6.0"
   },
   "peerDependencies": {
     "react": "^0.14.0 || ^15.0.0-0",
