+++ src/index.js
@@ -1,5 +1,4 @@
 export Model from './model';
 export Registry from './registry';
 export * as Attr from './attributes';
-export * from './redux/local';
-export * from './redux/remote';
+export * from './api';
