--- src/redux/local/constants.js
@@ -1,6 +1,0 @@
-export const LOCAL_CLEAR = 'AR_LOCAL_CLEAR';
-export const LOCAL_HYDRATE = 'AR_LOCAL_HYDRATE';
-export const LOCAL_READ = 'AR_LOCAL_READ';
-export const LOCAL_CREATE = 'AR_LOCAL_CREATE';
-export const LOCAL_UPDATE = 'AR_LOCAL_UPDATE';
-export const LOCAL_DELETE = 'AR_LOCAL_DELETE';
