--- src/redux/local/index.js
@@ -1,2 +1,0 @@
-export * from './actions';
-export localReducer from './reducer';
