--- src/redux/remote/index.js
@@ -1,2 +1,0 @@
-export * from './actions';
-export remoteReducer from './reducer';
